FIX: correct return type of read_pt

+++ control/src/pt.cc
@@ -13,7 +13,7 @@ float voltage_to_psi(const PtConfig& pt_config, float voltage) {
           (pt_config.voltage_range.second - pt_config.voltage_range.first));
 }
 
-uint16_t read_pt(Pt pt) {
+float read_pt(Pt pt) {
   const PtConfig& pt_config = get_pt_config(pt);
   float raw_voltage = pt_adc_read_raw_voltage(pt_config.cs, pt_config.channel);
